refactor(mir): I extract actor-state handle safety checks
